requirements.txt
pip~=21.0.1
nltk~=3.5
click~=7.1.2
tqdm~=4.59.0
requests~=2.25.1
regex~=2020.11.13
joblib~=1.0.1
six~=1.15.0
certifi~=2020.12.5
chardet~=4.0.0
urllib3~=1.26.3
psycopg2~=2.8.6
idna~=2.10
pymorphy2~=0.9.1
setuptools~=53.0.0
XlsxWriter~=1.3.7
apache-airflow[postgres]==2.0.1
colorlog~=4.7.2
cryptography~=3.4.6
numpy~=1.20.1
rich~=9.2.0
colorama~=0.4.4
psutil~=5.8.0
docutils~=0.16